% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listReferences.R
\name{listReferences}
\alias{listReferences}
\title{List available references}
\usage{
listReferences()
}
\arguments{
\item{name}{String containing the name of the reference dataset.}
}
\value{
For \code{listReferences}, a character vector containing the names of the available references.

For \code{listVersions}, a character vector containing the names of the available versions of the \code{name} reference.

For \code{fetchLatestVersion}, a string containing the name of the latest version.
}
\description{
List the available reference datasets and the associated versions in \pkg{celldex}.
}
\examples{
listReferences()
listVersions("immgen")
fetchLatestVersion("immgen")

}
\author{
Aaron Lun
}
