% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searchReferences.R
\name{searchReferences}
\alias{searchReferences}
\title{Search reference metadata}
\usage{
searchReferences(
  query,
  cache = cacheDirectory(),
  overwrite = FALSE,
  latest = TRUE
)
}
\arguments{
\item{query}{String or a \code{gypsum.search.object}, see Examples.}

\item{cache, overwrite}{Arguments to pass to \code{\link{fetchMetadataDatabase}}.}

\item{latest}{Whether to only consider the latest version of each dataset.}
}
\value{
A \linkS4class{DataFrame} where each row corresponds to a dataset, containing various columns of metadata.
Some columns may be lists to capture 1:many mappings.
}
\description{
Search for reference datasets of interest based on matching text in the associated metadata.
}
\details{
The returned DataFrame contains the usual suspects like the title and description for each dataset,
the number of rows and columns, the organisms and genome builds involved,
whether the dataset has any pre-computed reduced dimensions, and so on.
More details can be found in the Bioconductor metadata schema at \url{https://github.com/ArtifactDB/bioconductor-metadata-index}.
}
\examples{
searchReferences(defineTextQuery("immun\%", partial=TRUE))[,c("name", "title")]
searchReferences(defineTextQuery("10090", field="taxonomy_id"))[,c("name", "title")]
searchReferences(
   defineTextQuery("10090", field="taxonomy_id") &
   defineTextQuery("immun\%", partial=TRUE)
)[,c("name", "title")]

}
\seealso{
\code{\link{surveyReferences}}, to easily obtain a listing of all available datasets.
}
\author{
Aaron Lun
}
